Search alternates recursively as default for :Efoo
Adds a fallback for the navigation commands to search alternates for a
file matching the corresponding type. Alternates are searched
recursively, breadth-first. To keep the search sane and bounded, only
the first (most specific) set of alternates for a given file are
considered, and recursion is limited to a maximum depth.